refactor(audit): scan.js becomes files, stylesheets and templates modules
src/audit/scan.js (676 lines) mixed directory walking, ignore patterns,
git scoping, the Stylelint run, finding extraction, property recovery,
and Tailwind template scanning. Now three modules, each with one
responsibility and one direction: stylesheets and templates depend on
files for the type predicates; nothing depends back. The architecture
test caught the one cycle the mechanical split introduced (resolveScssSyntax
landed on the wrong side of the line) before any test ran.

Behaviour unchanged: 220 tests, benchmark check zero moved findings.
